Penyempurnaan error pada tahap membuat class controller sebelumnya

+++ core/src/Entity/Controller.java
@@ -5,29 +5,29 @@ import com.badlogic.gdx.Input;
 
 public class Controller {
 
-    int x,y;
-    Controller(int x, int y){
+    float x,y;
+    Controller(float x, float y){
         this.x=x;
         this.y=y;
     }
 
 
-    int naik(int y){
+    float naik(float y){
         y=+ 4;
         return y;
     }
 
-    int turun(int y){
+    float turun(float y){
         y=- 4;
         return y;
     }
 
-    int kanan(int x){
+    float kanan(float x){
         x=+4;
         return x;
     }
 
-    int kiri(int x){
+    float kiri(float x){
         x=-4;
         return x;
     }
